fix test

+++ katas/7kyu/Growth of population/solution.py
@@ -29,7 +29,7 @@
 def nb_year(p0, percent, aug, p):
     count = 0
     while p0 < p:
-        p0 += p0 * percent / 100 + aug
+        p0 += int(p0 * percent / 100 + aug)
         count += 1
 
     return count
